feature#2 패키지 이름 수정 및 타겟 이미지 조회 기능 추가
- 패키지 이름을 좀더 가시성 있고 통일성있게 소문자로 교체함

- 타겟 이미지에 thunbNail 항목을 추가해서 여러 항목을 받았더라도 한장만 썸네일로 추가 할수 있도록함

- 타겟정보 페이징 처리후 조회 할수 있도록 추가하고 타겟 정보와 연관된 이미지도 같이 반환 하도록 기능추가함
