+++ build.gradle
@@ -37,8 +37,8 @@ dependencies {
     testImplementation 'org.springframework.security:spring-security-test:5.6.2'
     testImplementation 'org.springframework.boot:spring-boot-starter-test:2.6.5'
     testCompileOnly 'org.projectlombok:lombok:1.18.22'
+    testAnnotationProcessor 'org.projectlombok:lombok:1.18.22'
     testImplementation 'com.h2database:h2:2.1.210'
-
 }
 
 tasks.named('test') {
