--- src/main/java/com/oldaim/fkbackend/DTO/ReissueDto.java
@@ -1,16 +1,0 @@
-package com.oldaim.fkbackend.DTO;
-
-import lombok.*;
-
-@ToString
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-@Setter
-public class ReissueDto {
-
-    private String userId;
-
-    private String refreshToken;
-
-}
