--- src/main/java/com/oldaim/fkbackend/Entity/BaseEntity.java
@@ -1,19 +1,0 @@
-package com.oldaim.fkbackend.Entity;
-
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedDate;
-
-import javax.persistence.MappedSuperclass;
-import java.time.LocalDateTime;
-
-
-
-@MappedSuperclass
-@Getter
-public abstract class BaseEntity {
-
-    @CreatedDate
-    private LocalDateTime localDateTime;
-
-
-}
