--- src/main/java/com/oldaim/fkbackend/Entity/EnumType/Auth.java
@@ -1,5 +1,0 @@
-package com.oldaim.fkbackend.Entity.EnumType;
-
-public enum Auth {
-    ROLE_USER,ROLE_ADMIN
-}
