--- src/main/java/com/oldaim/fkbackend/Entity/EnumType/ImageUsage.java
@@ -1,5 +1,0 @@
-package com.oldaim.fkbackend.Entity.EnumType;
-
-public enum ImageUsage {
-    TARGET,RETURN
-}
