--- src/main/java/com/oldaim/fkbackend/Repository/ImageRepository.java
@@ -1,9 +1,0 @@
-package com.oldaim.fkbackend.Repository;
-
-import com.oldaim.fkbackend.Entity.Image;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ImageRepository extends JpaRepository<Image,Long> {
-}
