--- src/main/java/com/oldaim/fkbackend/Repository/UserRepository.java
@@ -1,14 +1,0 @@
-package com.oldaim.fkbackend.Repository;
-
-import com.oldaim.fkbackend.Entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<User,Long> {
-    @Query("SELECT u FROM User u WHERE u.userId = ?1")
-    Optional<User> findByUserId(String username);
-}
