--- src/main/java/com/oldaim/fkbackend/Security/web/WebMvcConfig.java
@@ -1,17 +1,0 @@
-package com.oldaim.fkbackend.Security.web;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class WebMvcConfig implements WebMvcConfigurer {
-    @Override
-    public void addCorsMappings(CorsRegistry registry) {
-        registry.addMapping("/**")
-                .exposedHeaders("X-AUTH-TOKEN")
-                .allowCredentials(true)
-                .allowedMethods("POST","GET")
-                .allowedOrigins("http://localhost:3000");
-    }
-}
