+++ src/main/java/com/oldaim/fkbackend/controller/dto/ImagePathDto.java
@@ -1,0 +1,18 @@
+package com.oldaim.fkbackend.controller.dto;
+
+import lombok.*;
+
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+@Setter
+@Builder
+public class ImagePathDto {
+
+    private String filePath;
+
+    private String fileName;
+
+    private Long targetPk;
+
+}
