+++ src/main/java/com/oldaim/fkbackend/controller/dto/ReissueDto.java
@@ -1,0 +1,16 @@
+package com.oldaim.fkbackend.controller.dto;
+
+import lombok.*;
+
+@ToString
+@AllArgsConstructor
+@NoArgsConstructor
+@Getter
+@Setter
+public class ReissueDto {
+
+    private String userId;
+
+    private String refreshToken;
+
+}
