+++ src/main/java/com/oldaim/fkbackend/controller/dto/TargetInfoDto.java
@@ -1,0 +1,20 @@
+package com.oldaim.fkbackend.controller.dto;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class TargetInfoDto {
+
+  private Long targetPk;
+
+  private String userId;
+
+  private String personName;
+
+  private Long personAge;
+
+  private String characteristic;
+
+}
