+++ src/main/java/com/oldaim/fkbackend/entity/BaseEntity.java
@@ -1,0 +1,19 @@
+package com.oldaim.fkbackend.entity;
+
+import lombok.Getter;
+import org.springframework.data.annotation.CreatedDate;
+
+import javax.persistence.MappedSuperclass;
+import java.time.LocalDateTime;
+
+
+
+@MappedSuperclass
+@Getter
+public abstract class BaseEntity {
+
+    @CreatedDate
+    private LocalDateTime localDateTime;
+
+
+}
