+++ src/main/java/com/oldaim/fkbackend/entity/enumType/Auth.java
@@ -1,0 +1,5 @@
+package com.oldaim.fkbackend.entity.enumType;
+
+public enum Auth {
+    ROLE_USER,ROLE_ADMIN
+}
