+++ src/main/java/com/oldaim/fkbackend/entity/enumType/ImageUsage.java
@@ -1,0 +1,5 @@
+package com.oldaim.fkbackend.entity.enumType;
+
+public enum ImageUsage {
+    TARGET,RETURN
+}
