+++ src/main/java/com/oldaim/fkbackend/entity/enumType/ThumbNail.java
@@ -1,0 +1,5 @@
+package com.oldaim.fkbackend.entity.enumType;
+
+public enum ThumbNail {
+    Able , Unable
+}
