+++ src/main/java/com/oldaim/fkbackend/repository/UserRepository.java
@@ -1,0 +1,14 @@
+package com.oldaim.fkbackend.repository;
+
+import com.oldaim.fkbackend.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface UserRepository extends JpaRepository<User,Long> {
+    @Query("SELECT u FROM User u WHERE u.userId = ?1")
+    Optional<User> findByUserId(String username);
+}
