+++ src/main/java/com/oldaim/fkbackend/security/webConfig/WebMvcConfig.java
@@ -1,0 +1,17 @@
+package com.oldaim.fkbackend.security.webConfig;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class WebMvcConfig implements WebMvcConfigurer {
+    @Override
+    public void addCorsMappings(CorsRegistry registry) {
+        registry.addMapping("/**")
+                .exposedHeaders("X-AUTH-TOKEN")
+                .allowCredentials(true)
+                .allowedMethods("POST","GET")
+                .allowedOrigins("http://localhost:3000");
+    }
+}
