+++ src/test/resources/application.yml
@@ -17,3 +17,10 @@ spring:
         show_sql: true
         format_sql: true
 
+jwt:
+  secret: cZlsmdpaLksSAdsdmdaskjndasdas1j234214lasdasdasSnadaklsnd242adsadndkanskdskahfbj3123122k
+  access_valid_time: 360000000
+  refresh_valid_time: 10000000000000000
+
+upload:
+  path: "/Users/kimdonggyun/Desktop/bootpicture/"
